requirement.txt
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonify==0.5
mangum==0.19.0
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.2
pip-system-certs==4.0
PyJWT==2.10.1
pymongo==4.10.1
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
